Console UI Application
The application has been automated by using the console to take user input to run through various banking scenarios.
I'm thinking of adding a Bank object that which allow the application to create individual users and send money between different users. In the future unique features should be added to the Savings and Credit Card classes in order to take full advantage of the abstract Bank account class.